Test updates after disallowing Integer -> Contract conversion

+++ test/libsolidity/semanticTests/types/type_conversion_cleanup.sol
@@ -1,7 +1,7 @@
 contract Test {
-    function test() public returns (uint ret) { return uint(uint160(address(uint160(type(uint200).max)))); }
+    function test() public returns (uint ret) { return uint(uint160(address(uint160(uint128(type(uint200).max))))); }
 }
 // ====
 // compileViaYul: also
 // ----
-// test() -> 0xffffffffffffffffffffffffffffffffffffffff
+// test() -> 0xffffffffffffffffffffffffffffffff
